Relocate logic from AppHeader component to composables
